Hitting Tab selects current highlighted option
While navigating through native selects using the keyboard, one can start
typing the text of the desired option to select it, then hit the Tab key
to move to the next element with the desired option remaining selected.
This doesn't work with Chosen currently (the correct option is highlighted,
but it is not selected once the Tab key is hit.)

Now the current highlighted option in a non-multiple Chosen select is
selected when the Tab key is hit.

Resolves #104.